backend/requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
python-dotenv
google-generativeai
speechrecognition
pydub
ffmpeg-python
gtts
python-socketio[asgi]